+++ src/AnyStatus.Plugins/Widgets/Network/HTTP/OpenWebPage.cs
@@ -1,0 +1,9 @@
+﻿using AnyStatus.API;
+
+namespace AnyStatus
+{
+    public class OpenHttpWebPage : OpenWebPage<HttpStatus>
+    {
+        public OpenHttpWebPage(IProcessStarter ps) : base(ps) { }
+    }
+}
