+++ src/AnyStatus.Plugins/Widgets/Network/Ping/Ping.cs
@@ -1,0 +1,17 @@
+﻿using AnyStatus.API;
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+
+namespace AnyStatus
+{
+    [DisplayName("Ping")]
+    [DisplayColumn("Network")]
+    [Description("Test the reachability of a host")]
+    public class Ping : Widget, IHealthCheck, ISchedulable
+    {
+        [Required]
+        [Category("Ping")]
+        [Description("Host Name or IP Address")]
+        public string Host { get; set; }
+    }
+}
