+++ src/AnyStatus.Plugins/Widgets/Network/Port/NetworkProtocol.cs
@@ -1,0 +1,8 @@
+﻿namespace AnyStatus
+{
+    public enum NetworkProtocol
+    {
+        TCP,
+        UDP
+    }
+}
