+++ src/AnyStatus.Plugins/Widgets/Network/Speed/DownloadSpeed.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+
+namespace AnyStatus
+{
+    [DisplayName("Download Speed")]
+    [DisplayColumn("Network")]
+    [Description("Monitor network download speed.")]
+    public class DownloadSpeed : NetworkSpeed
+    {
+    }
+}
