+++ src/AnyStatus.Plugins/Widgets/Network/Speed/UploadSpeed.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
+
+namespace AnyStatus
+{
+    [DisplayName("Upload Speed")]
+    [DisplayColumn("Network")]
+    [Description("Monitor network upload speed.")]
+    public class UploadSpeed : NetworkSpeed
+    {
+    }
+}
