+++ src/AnyStatus.Plugins/Widgets/Pingdom/Pingdom.cs
